requirements.txt
fastapi
langchain_community
langchain_openai
uvicorn
neo4j==5.23.1
motor==3.5.1
datefinder==0.7.3
redis==5.0.8
flashRank
langchain-core
langchain-qdrant
python-dotenv==1.0.1
qdrant-client==1.12.1
langchain-qdrant
groq
pandas
BeautifulSoup4==4.12.3
nest_asyncio==1.6.0
selenium